Fixed meal creation
