+++ src/components/SignIn/signin.css
@@ -10,7 +10,7 @@ form {
     display: flex;
     flex-direction: column;
     line-height: 150%;
-    font-family: Trebuchet MS;
+    font-family: 'Roboto Slab', serif;
     font-size: 14px;
     color: #333333;
     padding: 20px;
@@ -34,7 +34,7 @@ input {
     font-weight: 500;
     font-size: 16px;
     line-height: 42px;
-    font-family: 'Helvetica', Arial, sans-serif;
+    font-family: 'Roboto Slab', serif;
     width: auto;
     white-space: nowrap;
     height: 42px;
