+++ src/components/app.css
@@ -5,7 +5,7 @@
 }
 
 body {
-    font-family: 'Roboto Slab', serif;;
+    font-family: 'Roboto Slab', serif;
     height: 110vh;
     background-color: #F3F4F6;
 }
